+++ Civi/DataProcessor/DataFlow/AbstractDataFlow.php
@@ -317,5 +317,20 @@ abstract class AbstractDataFlow {
     return $records;
   }
 
+  /**
+   * When an object is cloned, PHP 5 will perform a shallow copy of all of the
+   * object's properties. Any properties that are references to other
+   * variables, will remain references. Once the cloning is complete, if a
+   * __clone() method is defined, then the newly created object's __clone()
+   * method will be called, to allow any necessary properties that need to be
+   * changed. NOT CALLABLE DIRECTLY.
+   *
+   * @return void
+   * @link https://php.net/manual/en/language.oop5.cloning.php
+   */
+  public function __clone() {
+    $this->dataSpecification = clone $this->dataSpecification;
+  }
+
 
 }