+++ Civi/DataProcessor/DataFlow/CombinedDataFlow/CombinedSqlDataFlow.php
@@ -155,7 +155,10 @@ class CombinedSqlDataFlow extends SqlDataFlow implements MultipleSourceDataFlows
       $fields[] = $outputHandler->getAggregateFieldSpec()->getSqlGroupByStatement($this->getName());
     }
     foreach($this->sourceDataFlowDescriptions as $sourceDataFlowDescription) {
-      $fields = array_merge($fields, $sourceDataFlowDescription->getDataFlow()->getFieldsForGroupByStatement());
+      if (!$sourceDataFlowDescription->getDataFlow() instanceof SubqueryDataFlow) {
+        $fields = array_merge($fields, $sourceDataFlowDescription->getDataFlow()
+          ->getFieldsForGroupByStatement());
+      }
     }
     return $fields;
   }
@@ -213,8 +216,11 @@ class CombinedSqlDataFlow extends SqlDataFlow implements MultipleSourceDataFlows
       $clauses[] = $clause;
     }
     foreach($this->sourceDataFlowDescriptions as $sourceDataFlowDescription) {
-      if ($sourceDataFlowDescription->getDataFlow() instanceof SqlDataFlow) {
+      if ($sourceDataFlowDescription->getDataFlow() instanceof SqlTableDataFlow) {
         foreach($sourceDataFlowDescription->getDataFlow()->getWhereClauses() as $clause) {
+          if ($clause instanceof SqlDataFlow\WhereClauseInterface) {
+
+          }
           $clauses[] = $clause;
         }
       }
@@ -255,5 +261,23 @@ class CombinedSqlDataFlow extends SqlDataFlow implements MultipleSourceDataFlows
     return $this;
   }
 
+  /**
+   * When an object is cloned, PHP 5 will perform a shallow copy of all of the
+   * object's properties. Any properties that are references to other
+   * variables, will remain references. Once the cloning is complete, if a
+   * __clone() method is defined, then the newly created object's __clone()
+   * method will be called, to allow any necessary properties that need to be
+   * changed. NOT CALLABLE DIRECTLY.
+   *
+   * @return void
+   * @link https://php.net/manual/en/language.oop5.cloning.php
+   */
+  public function __clone() {
+    $sourceDataFlowDescriptions = array();
+    foreach($this->sourceDataFlowDescriptions as $sourceDataFlowDescription) {
+      $sourceDataFlowDescriptions[] = clone $sourceDataFlowDescription;
+    }
+    $this->sourceDataFlowDescriptions = $sourceDataFlowDescriptions;
+  }
 
 }